backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
supabase>=2.3.0

# Data Collection
pykrx>=1.0.45
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
pytrends>=4.9.2

# Analysis
pandas>=2.2.0
numpy>=1.26.3
ta>=0.11.0

# OpenAI
openai>=1.12.0

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Task Queue (optional)
# celery>=5.3.6
# redis>=5.0.1

# Scheduler (optional)
# apscheduler>=3.10.4

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.4
httpx>=0.26.0

# Utilities
python-dateutil>=2.8.2
